// src/api/generated/schemas/saveCombinationDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * StyleMind API
 * Documentación de la API de StyleMind
 * OpenAPI spec version: 1.0
 */
import type { CombinationItemDto } from './combinationItemDto';

export interface SaveCombinationDto {
  description: string;
  occasions: string[];
  name: string;
  isAIGenerated: boolean;
  combinationItems: CombinationItemDto[];
}